Perform test
