+++ .eslintrc.json
@@ -12,14 +12,14 @@
     "ecmaVersion": 2018,
     "sourceType": "module"
   },
-  // "extends": ["airbnb", "plugin:react/recommended"],
-  // "plugins": ["react"],
-  // "rules": {
-  //   "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
-  //   "react/react-in-jsx-scope": "off",
-  //   "import/no-unresolved": "off",
-  //   "no-shadow": "off"
-  // },
+  "extends": ["airbnb", "plugin:react/recommended"],
+  "plugins": ["react"],
+  "rules": {
+    "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
+    "react/react-in-jsx-scope": "off",
+    "import/no-unresolved": "off",
+    "no-shadow": "off"
+  },
   "ignorePatterns": [
     "dist/",
     "build/"
