--- README.old.md
@@ -1,1 +1,0 @@
-# crypto-market-app
