+++ src/Components/__test__/page.test.js
@@ -1,21 +1,24 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
-import store from '../../Redux/ConfigureStore'
-import HomePage from './../HomePage'
-import CoinPage from './../CoinPage'
+import store from '../../Redux/ConfigureStore';
+import HomePage from './../HomePage';
+import CoinDetailsRender from './../CoinDetailsRender';
 
-decribe('To take snapshot of all components', () => { 
+ 
+describe('To take snapshot of all components', () => { 
 test('To Test HomePage component if renders successfully!', () => {
-  const tree = renderer
+
+ const tree = renderer
     .create(<Provider store={store}><HomePage /></Provider>)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });
 
 test('To Test CoinPage component if renders successfully!', () => {
+  
   const tree = renderer
-    .create(<Provider store={store}><CoinPage /></Provider>)
+    .create(<Provider store={store}><CoinDetailsRender /></Provider>)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });
